// api/coreService/types/dto/AyahWithTranslationResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DtoAyahAudioFileResponse } from "./AyahAudioFileResponse.ts";
import type { DtoAyahTranslationResponse } from "./AyahTranslationResponse.ts";

export type DtoAyahWithTranslationResponse = {
    /**
     * @type object | undefined
    */
    audio?: DtoAyahAudioFileResponse;
    /**
     * @type string | undefined
    */
    createdAt?: string;
    /**
     * @type string | undefined
    */
    createdBy?: string;
    /**
     * @type integer | undefined
    */
    hizbNumber?: number;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type integer | undefined
    */
    juzNumber?: number;
    /**
     * @type integer | undefined
    */
    manzilNumber?: number;
    /**
     * @type integer | undefined
    */
    pageNumber?: number;
    /**
     * @type string | undefined
    */
    surahId?: string;
    /**
     * @type string | undefined
    */
    text?: string;
    /**
     * @type object | undefined
    */
    translation?: DtoAyahTranslationResponse;
    /**
     * @type string | undefined
    */
    updatedAt?: string;
    /**
     * @type string | undefined
    */
    updatedBy?: string;
};